LayoutTests/http/tests/contentextensions/reload-without-contentextensions-expected.txt

CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/contentextensions/reload-without-contentextensions.html from loading a resource from http://localhost:8000/resources/square128.png
PASS
